Converted to double to allow more precision. Added a function to compoare with double_epsilon when checking correctness to allow for double rounding error.

+++ d_main.h
@@ -1,1 +1,1 @@
-int d_discombobulate(float * energyGrid, atom *atoms, int dimX, int dimY, int dimZ, float gridSpacing,  int numAtoms, int which);
+int d_discombobulate(double * energyGrid, atom *atoms, int dimX, int dimY, int dimZ, float gridSpacing,  int numAtoms, int which);
